refactor: 앱 실행 시 음식점 목록을 초기화하는 로직을 커스텀 훅으로 분리
